+++ src/js/models/Lotto/Lotto.js
