+++ src/js/models/Lotto/LottoModel.js
