+++ src/js/models/LottoAppModel.js
